Revert "Android fb logic 3"
