+++ core/src/com/robinhood/game/RobinHood.java
@@ -1,9 +1,9 @@
 package com.robinhood.game;
 
 import com.badlogic.gdx.ApplicationAdapter;
-import com.robinhood.game.controller.Controller;
-import com.robinhood.game.view.MenuView;
-import com.robinhood.game.view.View;
+
+import controller.*;
+import view.*;
 
 public class RobinHood extends ApplicationAdapter {
