+++ core/src/model/Components.java
@@ -1,8 +1,10 @@
-package com.robinhood.game.model;
+package model;
 
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.scenes.scene2d.Actor;
+
+import java.awt.event.ComponentEvent;
 
 /*
      Component in ECS
