+++ core/src/view/GameOverView.java
@@ -1,4 +1,4 @@
-package com.robinhood.game.view;
+package view;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
@@ -8,13 +8,12 @@ import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
-import com.robinhood.game.controller.Controller;
-import com.robinhood.game.model.Model;
-import com.robinhood.game.view.interfaceObjects.Button;
 
 import java.util.List;
 
-
+import controller.Controller;
+import model.Model;
+import view.interfaceObjects.Button;
 
 public class GameOverView extends View {
