+++ core/src/view/MenuView.java
@@ -1,20 +1,20 @@
-package com.robinhood.game.view;
+package view;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
-import com.robinhood.game.controller.Controller;
-import com.robinhood.game.model.Model;
-import com.robinhood.game.model.SoundBar;
-import com.robinhood.game.view.interfaceObjects.Button;
 
-public class SettingsView extends View {
+import controller.Controller;
+
+import view.interfaceObjects.*;
+
+public class MenuView extends View {
 
     private final Controller controller;
 
-    public SettingsView(Controller cont, Model model) {
+    public MenuView(Controller cont) {
 
         this.controller = cont;
 
@@ -23,30 +23,41 @@ public class SettingsView extends View {
         super.stage = new Stage(new ScreenViewport());
         Gdx.input.setInputProcessor(super.stage);
 
-        Button menuButton = new Button("backbutton");
+        Button settingsButton = new Button("settings");
+        Button loadingButton = new Button("play");
+        Button exitButton = new Button("exit");
         Button robinHood = new Button("robinhoodpic");
-        SoundBar soundBar = model.getSoundBar();
-
-        // Checks if soundBar already has a clickListener, adds listener if not
-        if(!soundBar.getListener()){
-            soundBar.addListener(controller);
-        }
 
         // ClickListener: https://libgdx.badlogicgames.com/ci/nightlies/docs/api/com/badlogic/gdx/scenes/scene2d/utils/ClickListener.html
         // ClickListener triggered by user clicks on Button/Actor to call appropriate actions
-        menuButton.addListener(new ClickListener(){
+        settingsButton.addListener(new ClickListener(){
             @Override
             public void clicked(InputEvent event, float clickX, float clickY) {
-                System.out.println("TO MENU!");
-                controller.navigateTo("MENU");
+                System.out.println("TO SETTINGS!");
+                controller.navigateTo("SETTINGS");
+            }
+        });
+        loadingButton.addListener(new ClickListener(){
+            @Override
+            public void clicked(InputEvent event, float clickX, float clickY) {
+                System.out.println("TO GAME!");
+                controller.navigateTo("GAME");
+            }
+        });
+        exitButton.addListener(new ClickListener(){
+            @Override
+            public void clicked(InputEvent event, float clickX, float clickY) {
+                System.out.println("EXITING APP!");
+                controller.exitApplication();
             }
         });
 
         // Actor: https://libgdx.badlogicgames.com/ci/nightlies/docs/api/com/badlogic/gdx/scenes/scene2d/Actor.html
         // adds all the elements to this interface
-        super.stage.addActor(menuButton);
+        super.stage.addActor(settingsButton);
+        super.stage.addActor(loadingButton);
+        super.stage.addActor(exitButton);
         super.stage.addActor(robinHood);
-        super.stage.addActor(soundBar);
-    }
 
+        }
 }
