+++ core/src/view/SettingsView.java
@@ -1,4 +1,9 @@
-package com.robinhood.game.view;
+package view;
+
+import controller.Controller;
+import model.Model;
+import view.interfaceObjects.*;
+import model.SoundBar;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
@@ -6,14 +11,11 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import com.robinhood.game.controller.Controller;
-import com.robinhood.game.view.interfaceObjects.*;
-
-public class MenuView extends View {
+public class SettingsView extends View {
 
     private final Controller controller;
 
-    public MenuView(Controller cont) {
+    public SettingsView(Controller cont, Model model) {
 
         this.controller = cont;
 
@@ -22,41 +24,30 @@ public class MenuView extends View {
         super.stage = new Stage(new ScreenViewport());
         Gdx.input.setInputProcessor(super.stage);
 
-        Button settingsButton = new Button("settings");
-        Button loadingButton = new Button("play");
-        Button exitButton = new Button("exit");
+        Button menuButton = new Button("backbutton");
         Button robinHood = new Button("robinhoodpic");
+        SoundBar soundBar = model.getSoundBar();
+
+        // Checks if soundBar already has a clickListener, adds listener if not
+        if(!soundBar.getListener()){
+            soundBar.addListener(controller);
+        }
 
         // ClickListener: https://libgdx.badlogicgames.com/ci/nightlies/docs/api/com/badlogic/gdx/scenes/scene2d/utils/ClickListener.html
         // ClickListener triggered by user clicks on Button/Actor to call appropriate actions
-        settingsButton.addListener(new ClickListener(){
+        menuButton.addListener(new ClickListener(){
             @Override
             public void clicked(InputEvent event, float clickX, float clickY) {
-                System.out.println("TO SETTINGS!");
-                controller.navigateTo("SETTINGS");
-            }
-        });
-        loadingButton.addListener(new ClickListener(){
-            @Override
-            public void clicked(InputEvent event, float clickX, float clickY) {
-                System.out.println("TO GAME!");
-                controller.navigateTo("LOADING");
-            }
-        });
-        exitButton.addListener(new ClickListener(){
-            @Override
-            public void clicked(InputEvent event, float clickX, float clickY) {
-                System.out.println("EXITING APP!");
-                controller.exitApplication();
+                System.out.println("TO MENU!");
+                controller.navigateTo("MENU");
             }
         });
 
         // Actor: https://libgdx.badlogicgames.com/ci/nightlies/docs/api/com/badlogic/gdx/scenes/scene2d/Actor.html
         // adds all the elements to this interface
-        super.stage.addActor(settingsButton);
-        super.stage.addActor(loadingButton);
-        super.stage.addActor(exitButton);
+        super.stage.addActor(menuButton);
         super.stage.addActor(robinHood);
-
+        super.stage.addActor(soundBar);
     }
+
 }
