+++ core/src/view/interfaceObjects/DragIndicator.java
@@ -1,4 +1,4 @@
-package com.robinhood.game.view.interfaceObjects;
+package view.interfaceObjects;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
